[#3922] Ssh_authorized_key "Could not apply complete catalog"
If file backup is enabled, there are cases when attempts to backup a
user's ssh authorized keys can fail.  What happens is that when the
ssh_authorized_key type calls flush() itself calls super as
the desired user for the ssh_authorized_key.
This super call tries to create a backup directory for the
ssh_authorized_key target as the desired user which fails.
Calling super() beforehand ensures it can create the clientbucket
directory before backing it up.