[#3922] Ssh_authorized_key "Could not apply complete catalog"
If file backup is enabled, there are cases when attempts to backup a
user's ssh authorized keys can fail.  What happens is that when the
ssh_authorized_key type calls flush() on its parent which is
parsedfile.  When it tries to backup the file into the filebucket
the euid has already been set to the desired user so puppet
cannot create the filebucket directory.  The following
fix works around this by creating the directory as root